+++ hugegraph-core/src/main/java/com/baidu/hugegraph/backend/page/PageState.java
@@ -1,63 +1,56 @@
 /*
- * Copyright 2017 HugeGraph Authors
  *
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements. See the NOTICE file distributed with this
- * work for additional information regarding copyright ownership. The ASF
- * licenses this file to You under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ *  * Copyright 2017 HugeGraph Authors
+ *  *
+ *  * Licensed to the Apache Software Foundation (ASF) under one or more
+ *  * contributor license agreements. See the NOTICE file distributed with this
+ *  * work for additional information regarding copyright ownership. The ASF
+ *  * licenses this file to You under the Apache License, Version 2.0 (the
+ *  * "License"); you may not use this file except in compliance with the License.
+ *  * You may obtain a copy of the License at
+ *  *
+ *  *     http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ *  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ *  * License for the specific language governing permissions and limitations
+ *  * under the License.
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
  */
 
 package com.baidu.hugegraph.backend.page;
 
 import java.util.Base64;
-import java.util.Iterator;
 
-import com.baidu.hugegraph.HugeException;
+import com.baidu.hugegraph.backend.BackendException;
 import com.baidu.hugegraph.backend.serializer.BytesBuffer;
-import com.baidu.hugegraph.iterator.Metadatable;
 import com.baidu.hugegraph.util.Bytes;
 import com.baidu.hugegraph.util.E;
 
-public final class PageState {
+public class PageState {
 
-    public static final String PAGE = "page";
-    public static final String PAGE_NONE = "";
+    private final byte[] position;
+    private final int offset;
+    private final int total;
 
-    private int offset;
-    private String page;
-
-    public PageState(int offset, String page) {
-        E.checkArgument(offset >= 0, "The offset must be >= 0");
-        E.checkNotNull(page, "page");
+    public PageState(byte[] position, int offset, int total) {
+        E.checkNotNull(position, "position");
+        this.position = position;
         this.offset = offset;
-        this.page = page;
+        this.total = total;
     }
 
-    public void increase() {
-        this.offset++;
-        this.page = PAGE_NONE;
+    public byte[] position() {
+        return this.position;
     }
 
     public int offset() {
         return this.offset;
     }
 
-    public void page(String page) {
-        this.page = page;
-    }
-
-    public String page() {
-        return this.page;
+    public long total() {
+        return this.total;
     }
 
     @Override
@@ -66,9 +59,11 @@ public final class PageState {
     }
 
     public byte[] toBytes() {
-        BytesBuffer buffer = BytesBuffer.allocate(256);
+        int length = 2 + this.position.length + BytesBuffer.INT_LEN;
+        BytesBuffer buffer = BytesBuffer.allocate(length);
+        buffer.writeBytes(this.position);
         buffer.writeInt(this.offset);
-        buffer.writeString(this.page);
+        buffer.writeInt(this.total);
         return buffer.bytes();
     }
 
@@ -77,7 +72,7 @@ public final class PageState {
         try {
             bytes = Base64.getDecoder().decode(page);
         } catch (Exception e) {
-            throw new HugeException("Invalid page: '%s'", e, page);
+            throw new BackendException("Invalid page: '%s'", e, page);
         }
         return fromBytes(bytes);
     }
@@ -85,23 +80,15 @@ public final class PageState {
     public static PageState fromBytes(byte[] bytes) {
         if (bytes.length == 0) {
             // The first page
-            return new PageState(0, PAGE_NONE);
+            return new PageState(new byte[0], 0, 0);
         }
         try {
             BytesBuffer buffer = BytesBuffer.wrap(bytes);
-            int offset = buffer.readInt();
-            String page = buffer.readString();
-            return new PageState(offset, page);
+            return new PageState(buffer.readBytes(), buffer.readInt(),
+                                 buffer.readInt());
         } catch (Exception e) {
-            throw new HugeException("Invalid page: '0x%s'",
-                                    e, Bytes.toHex(bytes));
+            throw new BackendException("Invalid page: '0x%s'",
+                                       e, Bytes.toHex(bytes));
         }
-    }
-
-    public static String page(Iterator<?> iterator) {
-        E.checkState(iterator instanceof Metadatable,
-                     "Invalid paging iterator: %s", iterator.getClass());
-        Object page = ((Metadatable) iterator).metadata(PAGE);
-        return (String) page;
     }
 }
